# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Croatia eInvoice Adapter API",
    "description": "JSON-to-UBL generation and preflight validation using the official UBL 2.1 schema, EN 16931 rules, and Croatian CIUS/EXT rules. Readiness is exposed at /ready; paid production traffic must be blocked unless all engines are ready.",
    "contact": {
      "name": "FiscalRail Labs"
    },
    "license": {
      "name": "Proprietary API; third-party validation artefacts retain source licences"
    },
    "version": "0.3.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/ready": {
      "get": {
        "summary": "Readiness",
        "operationId": "readiness_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/info": {
      "get": {
        "summary": "Info",
        "operationId": "info_v1_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Info V1 Info Get"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hr/invoices/generate": {
      "post": {
        "summary": "Generate Invoice",
        "operationId": "generate_invoice_v1_hr_invoices_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hr/invoices/validate": {
      "post": {
        "summary": "Validate Invoice",
        "operationId": "validate_invoice_v1_hr_invoices_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/XmlValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/hr/invoices/preflight": {
      "post": {
        "summary": "Preflight Invoice",
        "operationId": "preflight_invoice_v1_hr_invoices_preflight_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Address": {
        "properties": {
          "street": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Street"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "City"
          },
          "postal_code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 2,
            "title": "Postal Code"
          },
          "country_code": {
            "type": "string",
            "const": "HR",
            "title": "Country Code",
            "default": "HR"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "street",
          "city",
          "postal_code"
        ],
        "title": "Address"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvoiceLine": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Description"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,20}|(?=[\\d.]{1,31}0*$)\\d{0,20}\\.\\d{0,10}0*$)"
              }
            ],
            "title": "Quantity"
          },
          "unit_price": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,20}|(?=[\\d.]{1,31}0*$)\\d{0,20}\\.\\d{0,10}0*$)"
              }
            ],
            "title": "Unit Price"
          },
          "unit_code": {
            "type": "string",
            "pattern": "^[A-Z0-9]{1,8}$",
            "title": "Unit Code",
            "default": "H87"
          },
          "vat_rate": {
            "type": "integer",
            "enum": [
              5,
              13,
              25
            ],
            "title": "Vat Rate",
            "default": 25
          },
          "kpd_code": {
            "type": "string",
            "title": "Kpd Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "description",
          "quantity",
          "unit_price",
          "kpd_code"
        ],
        "title": "InvoiceLine"
      },
      "InvoiceRequest": {
        "properties": {
          "invoice_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^\\S+$",
            "title": "Invoice Number"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "issue_time": {
            "type": "string",
            "format": "time",
            "title": "Issue Time"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "delivery_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Date"
          },
          "profile_id": {
            "type": "string",
            "maxLength": 200,
            "pattern": "^(P(?:[1-9]|1[0-2])|P99:.+)$",
            "title": "Profile Id",
            "default": "P1"
          },
          "currency": {
            "type": "string",
            "const": "EUR",
            "title": "Currency",
            "default": "EUR"
          },
          "supplier": {
            "$ref": "#/components/schemas/Party"
          },
          "customer": {
            "$ref": "#/components/schemas/Party"
          },
          "supplier_operator_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Supplier Operator Name",
            "default": "Automated system"
          },
          "payment_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Payment Id"
          },
          "supplier_iban": {
            "type": "string",
            "title": "Supplier Iban"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/InvoiceLine"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invoice_number",
          "issue_date",
          "issue_time",
          "due_date",
          "supplier",
          "customer",
          "payment_id",
          "supplier_iban",
          "lines"
        ],
        "title": "InvoiceRequest"
      },
      "Party": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "oib": {
            "type": "string",
            "title": "Oib"
          },
          "address": {
            "$ref": "#/components/schemas/Address"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "oib",
          "address"
        ],
        "title": "Party"
      },
      "PreflightResponse": {
        "properties": {
          "xml": {
            "type": "string",
            "title": "Xml"
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationResult"
          },
          "totals": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Totals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "xml",
          "validation",
          "totals"
        ],
        "title": "PreflightResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ValidationFinding": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "severity": {
            "type": "string",
            "enum": [
              "fatal",
              "error",
              "warning",
              "info"
            ],
            "title": "Severity"
          },
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "unknown"
          },
          "business_terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Business Terms"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "column": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rule_id",
          "severity",
          "path",
          "message"
        ],
        "title": "ValidationFinding"
      },
      "ValidationResult": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile",
            "default": "hr"
          },
          "profiles": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Profiles"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/ValidationFinding"
            },
            "type": "array",
            "title": "Findings"
          },
          "checks_run": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checks Run"
          },
          "production_ready": {
            "type": "boolean",
            "title": "Production Ready",
            "default": false
          },
          "engine": {
            "type": "string",
            "title": "Engine",
            "default": "fiscalrail-legacy/0.2.0"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valid",
          "findings",
          "checks_run"
        ],
        "title": "ValidationResult"
      },
      "XmlValidationRequest": {
        "properties": {
          "xml": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Xml"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "hr"
            ],
            "title": "Language",
            "default": "en"
          },
          "max_issues": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Max Issues",
            "default": 200
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "xml"
        ],
        "title": "XmlValidationRequest"
      }
    }
  }
}